xhs-video-mcp/pyproject.toml
[project]
name = "xhs-video-mcp"
version = "0.1.0"
description = "基于 FastMCP 的小红书视频生成 MCP 服务"
readme = "README.md"
authors = [
    { name = "luyike", email = "248618931@qq.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "moviepy>=2.1.2",
    "edge-tts==6.1.19",
    "requests>=2.31.0",
    "openai>=1.56.1",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xhs_video_mcp"]

[project.scripts]
xhs-video-mcp = "xhs_video_mcp.main:main"
